reportbug/utils.py: fix versioned provides recognition
get_package_info() so far did not support parsing package 'Provides:'
metadata lines where any version is specified. Fix this by recognizing
version restriction both for the package(s) we are looking for and the
packages that we are not interested in.

Since the recognized provides now may contain such a version
restriction and we really only want the package name, remove the
unneeded information before returning.