932383: utils.py: query dpkg status with unambiguous package name
Querying dpkg status for a binary package without arch qualifier does
not work when binaries for multiple architectures are installed. This
breaks local package installation detection in reportbug in such cases,
leading to unnecessary and confusing user prompts.

Fix this by using dpkg-query -W to append the arch qualifier if needed.
This means reportbug will use the native architecture of the system if
the user does not specify otherwise.