946363: recognize source packages when sending followup info
When reportbug's -N option is used and the user wants to submit followup
information, reportbug extracts the corresponding package from the
original report. This does not work well if the bug was reported against
a source package. Fix this by recognizing the 'src:' prefix.

The linux kernel is a special case due to the complexity of its
packaging structure.