debbugs.get_reports: add support for lists of packages
So far the 'package' argument to get_reports could be either a string
denominating a single package name, or a list of bug numbers. There is
a reason why more than one package name should be supported: Some
packages redirect bug reports to other packages through the Submit-As
feature of the bug control file. Therefore bug reports on a specific
package can potentially have been filed against different packages, and
it makes sense to query all of them.

Instead of introducing an additional submitas argument to get_reports,
this commit makes it possible to provide a list of package names to
get_reports, or even mixed lists of package names and bug numbers.