bin/reportbug: query both binary and source package by default
Since debbugs.get_reports() now supports lists of packages, we no longer
have to decide between either querying for the source or the binary
package bug lists. We can get a more comprehensive list of bugs by
querying for both at the same time.

The 'src' flag becomes unnecessary when we do this, so we don't need to
set it to True in this case.

For an even more comprehensive bug list, we should also consider any
Submit-As information from the bug control file.