debbugs.get_reports: remove duplicates from bug list
Since we now run multiple queries and merge the results to get a
comprehensive bug list, it is very likely that we end up with many
duplicates in the bug list. Removing the duplicates allows us (further
below) to determine the actual number of bugs as the length of the list.