utils.py: improve get_source_name
get_source_name is supposed to return the source package for the binary
package given as its argument. This sometimes returns the wrong source
package name. The reason is that `apt-cache showsrc` can return a list
of packages, and the one we are looking for is not always listed first.

To fix this, switch to `apt-cache show`, which only looks for binary
packages and only returns a single answer.

Further modifications are needed after this: The regular expression now
needs to match on 'Source:'. That is also only present if the source
package name differs from the given binary package name, so we need to
handle this case in addition to the no-answer case.