gtk_ui/handle_bts_query: Use QuertBTSError instead of NoBugs
When `debbugs.get_reports` throws an exception, this clearly
indicates a problem with querying the BTS. 'NoBugs' just
suggests that debbugs.get_reports did not return any bugs,
which is also technically true, but why not use the more
specific exception here?

This commit also fixes the problem that the `error_dialog`
gtk_ui helper function cannot be called here - it can only
be called from within the ui context, while this function
is actually running in reportbug context.

To keep the error message (slightly extended as copied from
the text ui version of `ui.handle_bts_query`) and be able
to show it to the user, add it to the Exception we are throwing.