text_ui/handle_bts_query: Use QuertBTSError
This aligns behavior of the text ui with the new behavior
of the gtk ui. This implies dropping the user question
(`select_option`) here since we cannot do anything
similar in the gtk ui.

We can move the user question into reportbug itself instead;
this is done in a followup commit.