bin/reportbug: Improved QuertBTSError handling
Previously only the text ui asked the user if they
wanted to continue if the BTS could not be reached.
With more unified usage of QuertBTSError across
the UIs in the previous commits, we can ask this
question in all UIs now.

Also show a little more information to point at
the underlying issue.