961955: urlutils.py, checkversions.py: use proxy and timeout
Querying madison for newer versions can hang for a long time if the
query is run without a timeout parameter. The reportbug documentation
says reportbug is using a configurable timeout parameter, so do it.

In reportbug/checkversions.py: pass proxy and timeout parameters to
`urlutils.open_url()` when querying madison for newer version

In reportbug/urlutils.py: Actually use the parameters in `urlopen()`
that are passed from `open_url()`