reportbug/utils.py: add mua support via xdg-email
xdg-email will automatically (try to) pick the user's preferred
email software; this should work in most desktop environments.
If it does not work, try setting the MAILER env var to point
to an email software that understands mailto URIs.

We need to pass everything to xdg-email as a mailto URL already,
since this lets us avoid shell escaping issues with content
that we do not control.

Closes: #532150
Closes: #788045
Closes: #865792