utils.Mua.send: return commandline instead of calling ui.system
The trouble with using ui.system() in Mua.send() is that it always
resolves to being the one from the text ui, even when we are running the
GTK interface. This does not work. To avoid the problem, do not call
ui.system() in Mua.send(). Just return the command line to be used by
the caller, and call ui.system() from submit.send_report() instead.