Use X-Debbugs-CC as pseudoheader instead of header
Debbugs understands this, see #179340 or #503996.

Using a pseudoheader makes it immediately obvious to recipients why they
receive the email. As a pseudoheader it also won't get lost when the
submitter passes the message to their preferred MUA before sending.

Sending X-Debbugs-Cc as pseudoheaders by default also means that we do
not need an extra dialog for this in handle_editing() in bin/reportbug.
Users can just enter such a pseudoheader while editing the message if
they want. Contrary to the CCS global array method used previously, this
actually also works when passing the message to a MUA directly from
handle_editing.

For pseudoheaders we need to be careful with the capitalization unless
utils.cleanup_msg() gets changed. It must be X-Debbugs-Cc, not -CC.