bin/reportbug: handle_editing: capitalize "print ..." menu option
All other options start with a capital letter, and so should this one.
This is particularly nice since the GTK ui checks the option strings to
match this one, which fails because it looks for "Print ...". So this
reintroduces the "Display message in text view" option in the GTK ui
version of this dialog.