mailer.py,submit.py: only run MUA in terminal if needed
The problem with always using ui.system() to call the Mua in
reportbug/submit.py is the GTK ui variant of this function, which seems
to try to kill child processes that are still running after the main
command has returned (and the terminal has been closed). Regardless of
whether this killing always works or not, it definitely makes using the
xdg-email Mua practically unusable.

The solution is to avoid spawning a terminal for Mua invocation if we
can (i.e., for all MUAs that don't need a terminal).

This commit
 - adds a `needs_terminal' property to the Mua class in
   reportbug/mailer.py, and sets the appropriate value for the supported
   MUAs;
 - uses this property in reportbug/submit.py to recognize if using
   ui.system() is needed. We better use subprocess.run() if not.

Closes: #964846