reportbug/bugreport.py: add second newline to empty body
There should always be one empty line between the pseudoheaders and the
main bug message text. With the second newline, also expert users of the
GTK interface can start typing their message directly in the message
editor, and do not need to start by inserting another newline.

Non-expert users are not affected because they do not start from an
empty message, but are given a nice message template with instructions.