bin/reportbug, reportbug/ui/*: unify yes_no question mark handling
ui.yes_no was called inconsistently with and without question mark in
the actual questions; about half of the questions had it and the other
half did not.  Unify these to always include it.  Then we don't have to
add special code to the gtk and urwid UI variants of the function to add
it only where needed.

Instead only the text UI version of the function should get special
code, because it wants to include keys to the available answers at the
end of the question like this: `[y|N|q|?]?`