bin/reportbug: add --resume-saved option
The temporary files where reportbug saves unsubmitted reports so far
couldn't be handled very well: While they are textfiles, they are only
human-readable if the original text did not contain non-ascii
characters; otherwise they are base64-encoded. Instructing the user to
remove the header lines and include the text via `--body-file` or
`--include` is not adequate advice in that case. The only reliable way I
know to deal with these outfiles is to use `mutt -H` on them, but that
is not always available where reportbug is used.

This commit solves the problem by introducing a new `--resume-saved`
option, which actually parses and decodes the saved report given to it.
It does not need prior editing of the file and works reliably with
encoded text. The reportbug prompts for package, version, subject, tags
etc are skipped since all of this should be given in the report already
and can be edited afterwards anyway.

Message attachments already stored in the tempfile are ignored. Users
will need to reattach attachments in order to include them in the
submitted message.

Closes: #328531, #702107