reportbug/debbugs.py: include and pre-select default in menus
In the GTK ui we sometimes end up with a non-default item pre-selected,
and users have difficulty figuring out that ctrl-click is necessary to
unselect it again. To avoid this situation for the suite selection for
binNMU and package removal bugs, include the default option (unstable)
in the corresponding menus and pre-select it.

Closes: #855172