conf/reportbug.conf, man/reportbug.conf.5: do not mention outfile
Putting reportbug's output filename in the configuration file is a bad
idea - it can lead to data loss when people accidentally overwrite
previous reports that have not been submitted yet. Supporting this as a
command line flag is okay, but this really should not exist as a config
file setting.

Now people might still be using it so I'm not sure we can just drop
support for this "feature". So let's start with no longer advertising
it. When we later want to drop support, we should introduce a warning
message whenever reportbug finds this setting in the config file.

The motivation for introducing this at the time was to have an easy
option to work around the limited MUA support in reportbug. This is no
longer an issue since MUA support is much better these days.