reportbug/submit.py: drop newlines from final_message text
Since newlines in long_message and final_message texts are now
consistently preserved in the three UIs, we should not include them in
message texts when we don't want them. There is only one message text
that included newlines implicitly by being written as a triple-quoted
string: convert it to single-quoted to drop the unwanted newlines.