reportbug/debbugs.py: handle installation and upgrade reports
Report templates for installation and upgrade reports have existed
outside reportbug for a long time. Since we already include templates
for various other virtual packages, we can incorporate these, too.

The installation-report package exists to assist with installation
reports by collecting system information with a bug script. The fact
that it also produced the template lead to issues with reportbug's GTK
interface, where the system information part of the message cannot be
edited interactively by the user. Solving this problem using the
approach implemented in this commit implies that the installation-report
bug script should be changed to no longer output the report template.

Closes: #931438