checkversions.get_versions_available(): distinguish src and bin pkg versions
Matching package names of a source and a binary package do not imply
that the binary is built from that source, and indeed there are packages
in Debian where this is not the case.

When looking up available versions of a binary package, skip matching
source packages; only consider binary package versions.

Closes: #873493