bin/reportbug: improve version lookup for source packages
Use the new utils.get_source_version() instead of walking through binary
packages to obtain the version. This allows to get the version earlier
and avoid checking binary status altogether for source packages.

Also avoid the use of self.options.pkgversion and self.options.source in
the condition for the package version prompt later, because they may no
longer be the same as pkgversion and issource at this stage. This
sometimes resulted in unnecessary package version prompts.

Closes: #984904