bin/reportbug: adapt to new version of utils.get_source_package
get_source_package now includes the actual source package in the
returned list, so we no longer have to add it here. We only added
it here if it did not match the name of a binary package anyway,
but it is better to always include it.

The list of binary packages is already sorted and was before already.
Drop the redundant .sort().

Drop a redundant check for an empty list of packages.

If the list of packages has only one entry, we can use that also
in non-interactive mode.

When evaluating the user's package selection from the menu and it
is a source package, directly get the version as well.